requirements.txt
Flask
flask-cors
pymongo
Pillow
pytesseract
tensorflow
numpy
opencv-python
scikit-learn
